requirements.txt
requests
selenium
webdriver-manager
python-dotenv